+++ base/nvti.c
@@ -44,10 +44,15 @@
  * management etc.
  */
 
+/* For strptime in time.h. */
+#undef _XOPEN_SOURCE
+#define _XOPEN_SOURCE
 #include "nvti.h"
 
-#include <stdio.h>
-#include <string.h> // for strcmp
+#include <stdio.h>   // for sscanf
+#include <string.h>  // for strcmp
+#include <strings.h> // for strcasecmp
+#include <time.h>    // for strptime
 
 #undef G_LOG_DOMAIN
 #define G_LOG_DOMAIN "lib  nvti"
@@ -152,6 +157,109 @@ const gchar *
 vtref_text (const vtref_t *r)
 {
   return (r ? r->ref_text : NULL);
+}
+
+/* Support function for timestamps */
+
+/**
+ * @brief Try convert an NVT tag time string into epoch time
+ *        or return 0 upon parse errors.
+ *
+ * @param[in]   str_time Time stamp as string in one of the forms used in NVTs.
+ *
+ * @return Time as seconds since the epoch.
+ */
+static time_t
+parse_nvt_timestamp (const gchar *str_time)
+{
+  time_t epoch_time;
+  int offset;
+  struct tm tm;
+
+  if ((strcmp ((char *) str_time, "") == 0)
+      || (strcmp ((char *) str_time, "$Date: $") == 0)
+      || (strcmp ((char *) str_time, "$Date$") == 0)
+      || (strcmp ((char *) str_time, "$Date:$") == 0)
+      || (strcmp ((char *) str_time, "$Date") == 0)
+      || (strcmp ((char *) str_time, "$$") == 0))
+    {
+      return 0;
+    }
+
+  /* Parse the time. */
+
+  /* 2011-08-09 08:20:34 +0200 (Tue, 09 Aug 2011) */
+  /* $Date: 2012-02-17 16:05:26 +0100 (Fr, 17. Feb 2012) $ */
+  /* $Date: Fri, 11 Nov 2011 14:42:28 +0100 $ */
+  memset (&tm, 0, sizeof (struct tm));
+  if (strptime ((char *) str_time, "%F %T %z", &tm) == NULL)
+    {
+      memset (&tm, 0, sizeof (struct tm));
+      if (strptime ((char *) str_time, "$Date: %F %T %z", &tm) == NULL)
+        {
+          memset (&tm, 0, sizeof (struct tm));
+          if (strptime ((char *) str_time, "%a %b %d %T %Y %z", &tm) == NULL)
+            {
+              memset (&tm, 0, sizeof (struct tm));
+              if (strptime ((char *) str_time, "$Date: %a, %d %b %Y %T %z", &tm)
+                  == NULL)
+                {
+                  memset (&tm, 0, sizeof (struct tm));
+                  if (strptime ((char *) str_time, "$Date: %a %b %d %T %Y %z",
+                                &tm)
+                      == NULL)
+                    {
+                      g_warning ("%s: Failed to parse time: %s", __FUNCTION__,
+                                 str_time);
+                      return 0;
+                    }
+                }
+            }
+        }
+    }
+  epoch_time = mktime (&tm);
+  if (epoch_time == -1)
+    {
+      g_warning ("%s: Failed to make time: %s", __FUNCTION__, str_time);
+      return 0;
+    }
+
+  /* Get the timezone offset from the str_time. */
+
+  if ((sscanf ((char *) str_time, "%*u-%*u-%*u %*u:%*u:%*u %d%*[^]]", &offset)
+       != 1)
+      && (sscanf ((char *) str_time, "$Date: %*u-%*u-%*u %*u:%*u:%*u %d%*[^]]",
+                  &offset)
+          != 1)
+      && (sscanf ((char *) str_time, "%*s %*s %*s %*u:%*u:%*u %*u %d%*[^]]",
+                  &offset)
+          != 1)
+      && (sscanf ((char *) str_time,
+                  "$Date: %*s %*s %*s %*u %*u:%*u:%*u %d%*[^]]", &offset)
+          != 1)
+      && (sscanf ((char *) str_time,
+                  "$Date: %*s %*s %*s %*u:%*u:%*u %*u %d%*[^]]", &offset)
+          != 1))
+    {
+      g_warning ("%s: Failed to parse timezone offset: %s", __FUNCTION__,
+                 str_time);
+      return 0;
+    }
+
+  /* Use the offset to convert to UTC. */
+
+  if (offset < 0)
+    {
+      epoch_time += ((-offset) / 100) * 60 * 60;
+      epoch_time += ((-offset) % 100) * 60;
+    }
+  else if (offset > 0)
+    {
+      epoch_time -= (offset / 100) * 60 * 60;
+      epoch_time -= (offset % 100) * 60;
+    }
+
+  return epoch_time;
 }
 
 /* VT Information */
@@ -1059,6 +1167,10 @@ nvti_set_solution_type (nvti_t *n, const gchar *solution_type)
 
 /**
  * @brief Add a tag to the NVT tags.
+ *        The tag names "last_modification" and "creation_date" are
+ *        treated special: The value is expected to be a timestamp
+ *        and it is being converted to seconds since epoch before
+ *        added as a tag value.
  *
  * @param n     The NVT Info structure.
  *
@@ -1071,6 +1183,8 @@ nvti_set_solution_type (nvti_t *n, const gchar *solution_type)
 int
 nvti_add_tag (nvti_t *n, const gchar *name, const gchar *value)
 {
+  gchar *newvalue = NULL;
+
   if (!n)
     return (-1);
 
@@ -1080,16 +1194,31 @@ nvti_add_tag (nvti_t *n, const gchar *name, const gchar *value)
   if (!value || !value[0])
     return (-3);
 
+  if (!strcmp (name, "last_modification"))
+    {
+      nvti_set_modification_time (n, parse_nvt_timestamp (value));
+      newvalue = g_strdup_printf ("%i", (int) nvti_modification_time (n));
+    }
+  else if (!strcmp (name, "creation_date"))
+    {
+      nvti_set_creation_time (n, parse_nvt_timestamp (value));
+      newvalue = g_strdup_printf ("%i", (int) nvti_creation_time (n));
+    }
+
   if (n->tag)
     {
       gchar *newtag;
 
-      newtag = g_strconcat (n->tag, "|", name, "=", value, NULL);
+      newtag =
+        g_strconcat (n->tag, "|", name, "=", newvalue ? newvalue : value, NULL);
       g_free (n->tag);
       n->tag = newtag;
     }
   else
-    n->tag = g_strconcat (name, "=", value, NULL);
+    n->tag = g_strconcat (name, "=", newvalue ? newvalue : value, NULL);
+
+  if (newvalue)
+    g_free (newvalue);
 
   return (0);
 }